requirements.txt
djangorestframework-csv

